+++ firehose-tracer/src/pb/pb.rs
@@ -23,4 +23,11 @@ pub mod sf {
             }
         }
     }
+    pub mod firehose {
+        // @@protoc_insertion_point(attribute:sf.firehose.v2)
+        pub mod v2 {
+            include!("sf.firehose.v2.rs");
+            // @@protoc_insertion_point(sf.firehose.v2)
+        }
+    }
 }